requirements.txt
datasets
numpy
collections
tqdm
matplotlib
fastapi 
uvicorn
torchtext
hydra-core
omegaconf
rich